Delete language facet when querying apache solr 8
Creating a new (file based) cores with apache solr resulted in having
a language facet along with the desired one. I am not sure if this is a
feature of the new solr version or I've fudged something up in the new
schema. To make this work meanwhile, I've decided to delete occurring
language facets.